Notes/pyproject.toml
[project]
name = "visualizers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "django>=6.0.1",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "markdown>=3.10",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pymdown-extensions>=10.20",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]
